+++ rbx_binary/CHANGELOG.md
@@ -1,6 +1,9 @@
 # rbx_binary Changelog
 
 ## Unreleased
+* Added support for `SecurityCapabilities` values. ([#359])
+
+[#359]: https://github.com/rojo-rbx/rbx-dom/pull/359
 
 ## 0.7.1 (2023-08-09)
 * Added support for `UniqueId` values. ([#271])